Revert "Merge pull request #4 from mashiox/master"
This reverts commit 50eb4598dc8376ad2583899c3454075bf524ab54, reversing
changes made to 3d0248a9b64cfcc971a96f53c8913b48825274ac.
